etl/requirements.txt
# Azure SDKs
azure-identity>=1.16.0
azure-cosmos>=4.6.0
azure-storage-blob>=12.20.0

# ETL helpers
requests>=2.32.0
